// old/src/app/models/external/order/nit-nombre-order-param.ts
export interface NitNombreOrderParam{
    tipoFacturacion?: number;
    ordenId?:number;
    nit?:string;
    idTipoDocumento?: number;
    tipoDocumento?: string;
    documento?: string;
    complemento?: string;
    nombre?: string;
    correo?: string;
    telefono?: string;
    cardNumber?: string;
}